requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
boto3==1.23.2
botocore==1.26.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
logzero==1.7.0
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
pandas==1.4.2
peewee==3.14.10
pg==0.1
Pillow==9.1.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
random-user-agent==1.0.1
regex==2022.4.24
requests==2.27.1
rollbar==0.16.2
s3transfer==0.5.2
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.7.0
Werkzeug==2.1.2
wsproto==1.1.0
xlrd==1.2.0
zipp==3.8.0
